requirements.txt
flask
python-dotenv
together
sentence-transformers
chromadb
uvicorn
